Add video feature to landing page (#10)
This pull request adds a video feature tot the landing page

**Code Changes:**
- Implemented a `media-container` class to handle both video and image
display in the `index.profile.html` file.
- Updated the CSS to make the video fit the container, in a 16:9 aspect
ratio.
- Added logic to automatically show a video with a thumbnail image when
both are available, show only the video when an image isn't provided,
and show only the image when a video isn't available.

- Included a `video` tag with controls and a thumbnail image to show
before the video plays. The video container has a `max-width` of 100%,
so it can resize across different screen sizes. Added `margin-top` for
improved spacing.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **New Features**
- Added video playback functionality to user profiles. Images are
replaced with a video player if `videoUrl` is provided.
- **Chores**
- Updated `.gitignore` to exclude the virtual environment directory from
version control.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: savas <stefania.sava@ait.ac.at>